Simplifying the fix for #80 when a server is unreachable: simpler refactoring.
This commit simplifies the handling of the problem by removing the new
properties added to the Registry domain object. Instead, adding an render
object in the Repostory controller for the UI.

One of the driving behavior was that the previous commit was not working
properly, where the value of "unreachable" was being cached.

Finally, using the alert message documented in the used Theme used by
the application "Journal" at http://bootswatch.com/journal/.

*	modified:   grails-app/controllers/docker/registry/web/RepositoryController.groovy
- Adding a new data type to map the registry to the current search status.
- Adding the status call and removing the call to the removed method.
- Adding the map "reachable" to the render.

*	modified:   grails-app/domain/docker/registry/web/Registry.groovy
- Removing the transient property from the Domain object and the transients list.
- Removing the unecessary method.

*	modified:   grails-app/views/repository/index.gsp
- Using the alert message only if the server is NOT reachable.