[bugfix] Refactoring RegistrySpec test cases for the fix #82
Fixing the test cases for the RegistrySpec in regards to the default
port number from URLs. The previous push was still broken. I was
confused with the toURL method, which has the same name as the Java's
URL class.

So this one fixes #82 and makes it ready for merge! Here's the test
execution.

$ grails test-app docker.registry.ui.Registry
Configuring Spring Security Core ...
... finished configuring Spring Security Core
2014-09-19 11:29:26,345 [main] INFO  conf.BootStrap  - Checking for registries in system env
| Tests PASSED - view reports in /home/mdesales/dev/github/docker-registry-web/target/test-reports